+++ crates/mimic/src/db/store/data.rs
@@ -2,7 +2,6 @@ use crate::{
     core::{
         Key,
         traits::{EntityKind, Storable},
-        types::Principal,
     },
     db::hasher::xx_hash_u64,
     ic::structures::{BTreeMap, DefaultMemory, storable::Bound},
@@ -70,6 +69,8 @@ pub struct DataKey {
 }
 
 impl DataKey {
+    pub const STORABLE_MAX_SIZE: u32 = 29;
+
     #[must_use]
     pub fn new(path: &str, key: impl Into<Key>) -> Self {
         Self {
@@ -87,16 +88,16 @@ impl DataKey {
     }
 
     #[must_use]
-    pub fn max() -> Self {
-        Self {
-            entity_id: u64::MAX,
-            key: Principal::MAX.into(),
-        }
+    pub const fn key(&self) -> Key {
+        self.key
     }
 
     #[must_use]
-    pub const fn key(&self) -> Key {
-        self.key
+    pub fn max_self() -> Self {
+        Self {
+            entity_id: u64::MAX,
+            key: Key::MAX,
+        }
     }
 }
 
@@ -112,9 +113,7 @@ impl From<DataKey> for Key {
     }
 }
 
-// current max serialized size is 66 bytes
-// making it up to 72 to add a buffer
-impl_storable_bounded!(DataKey, 72, false);
+impl_storable_bounded!(DataKey, Self::STORABLE_MAX_SIZE, false);
 
 ///
 /// DataKeyRange
@@ -216,17 +215,12 @@ mod tests {
     use crate::core::types::Ulid;
 
     #[test]
-    fn max_size_is_66_bytes() {
-        let key = DataKey::max();
-        let bytes = key.to_bytes();
-        println!("Storable serialized size: {}", bytes.len());
+    fn data_key_max_size_is_bounded() {
+        let data_key = Key::max_self();
+        let size = Storable::to_bytes(&data_key).len() as u32;
 
-        assert_eq!(bytes.len(), 66);
-    }
-
-    #[test]
-    fn data_key_is_48_bytes() {
-        assert_eq!(std::mem::size_of::<DataKey>(), 48);
+        println!("max serialized size = {size}");
+        assert!(size <= Key::STORABLE_MAX_SIZE);
     }
 
     #[test]